Restore KGD.html report as per legacy behaviour
This is an important file GenomNZ staff expect to be able to find.
