+++ src/agr/gbs_prism/redun/__init__.py
@@ -3,7 +3,7 @@
 from .stage1 import run_stage1, Stage1Output
 from .stage2 import run_stage2, Stage2Output
 from .stage3 import run_stage3, Stage3Output
-from .reports import create_reports
+from .reports.reports import create_reports
 from .warehouse import warehouse
 
 __all__ = [
